NXP-28863: extract to a shared element wrapped by the existing ones
