+++ i18n/messages.json
@@ -184,6 +184,7 @@
   "audit.performedAction": "Performed Action",
   "audit.selectEventCategory": "Select event category",
   "audit.selectEventTypes": "Select event actions",
+  "audit.state": "State",
   "audit.to": "To",
   "audit.username": "Username",
   "audit.usernamePlaceholder": "Filter by username",
@@ -438,12 +439,6 @@
   "documentHistory.empty": "No results",
   "documentHistory.filter.after": "After",
   "documentHistory.filter.before": "Before",
-  "documentHistory.filter.eventCategory": "Event Category",
-  "documentHistory.filter.eventType": "Performed Actions",
-  "documentHistory.filter.selectEventCategory": "Select event category",
-  "documentHistory.filter.selectEventTypes": "Select event actions",
-  "documentHistory.filter.username": "Username",
-  "documentHistory.filter.usernamePlaceholder": "Filter by username",
   "documentHistory.performedAction": "Performed Action",
   "documentHistory.state": "State",
   "documentHistory.username": "Username",
